change compute_isi_computation to take data arg
